% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_zscore.R
\name{calc_revzscore}
\alias{calc_revzscore}
\title{\code{calc_zscore} Returns the raw value of a z score given the non-standardized vector v.}
\usage{
calc_revzscore(z, v)
}
\arguments{
\item{z}{a numeric value}

\item{v}{a numeric vector}
}
\description{
\code{calc_zscore} Returns the raw value of a z score given the non-standardized vector v.
}
