% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_zscore.R
\name{calc_zscore}
\alias{calc_zscore}
\title{\code{calc_zscore} Returns the z score of a raw value given the non-standardized vector v.}
\usage{
calc_zscore(z, v)
}
\arguments{
\item{z}{a numeric value}

\item{v}{a numeric vector}
}
\description{
\code{calc_zscore} Returns the z score of a raw value given the non-standardized vector v.
}
