% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIR.r
\name{PIRcalc}
\alias{PIRcalc}
\title{Calculate Poverty Income Ratio}
\usage{
PIRcalc(income, size, children, year = 2018)
}
\arguments{
\item{income}{vector of total household income}

\item{size}{vector of total family size (including children)}

\item{year}{vector of number of children in the family.}
}
\description{
This function calculates the ratio of total family income given number of adults and children in a family.
You must provide the year of the poverty thresholds you want to use, which are presently hard coded (it would be great to move them to a package data folder later).
}
