% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDI2.r
\name{score_BDI2}
\alias{score_BDI2}
\title{Score the Beck Depression Inventory (v2)}
\usage{
score_BDI2(meas)
}
\description{
This is a scoring program for the Beck Depression Inventory (v2).
It accepts a data frame with 21 columns corresponding to responses
on the measure.
}
\details{
\code{score_BDI2} Returns a symptom total.
}
