% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQAQ.R
\name{score_BQAQ}
\alias{score_BQAQ}
\title{Score the Buss-Perry Aggression Questionnaire (BQAQ)}
\usage{
score_BQAQ(meas)
}
\arguments{
\item{meas}{n×29 vector of reported responses}
}
\description{
This is a scoring program for the BQAQ
}
\details{
code{score_BQAQ} Returns cluster scoles for BQAQ subscales and total score.
}
