% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAPS-V.R
\name{score_CAPS}
\alias{score_CAPS}
\title{This is a scoring program for the CAPS-V, using rules from our MT dataset.}
\usage{
score_CAPS(
  crita,
  life_frequency,
  curr_frequency,
  life_intens = 0,
  curr_intens = 0,
  nacorrect = FALSE,
  correct.errors = FALSE,
  cluster.scores = FALSE
)
}
\arguments{
\item{crita}{n×1 vector of criterion A (qualifying trauma)}

\item{life_frequency}{n×25 vector of reported lifetime frequency.}

\item{curr_frequency}{n×25 vector of reported current frequency.}

\item{correct.errors}{should the program correct common
interview errors. Currently this means if lifetime score for a
symtom is lower than current, lifetime will be overwritten.}

\item{cluster.scores}{should the program print cluster scores.}

\item{life_intensity}{n×20 vector of reported lifetime intensity -
defaults to 0 for cases where intensity not reported (eg parent
report).}

\item{curr_intensity}{n×20 vector of reported current intensity -
defaults to 0 for cases where intensity not reported (eg parent
report).}

\item{skipped.is.na}{are items which have been properly skipped
(eg. all items following absence of crit A) coded as NA? If so,
the program will substitute appropriately with zeroes before
scoring. Otherwise it will attempt to score with standard R
handling of missing values.}
}
\description{
\code{score_CAPS} Returns current and lifetime PTSD symptom counts and
diagnosis codes given appropriate input. Typically input is given as
column subsets of a larger dataframe. All inputs must be numeric.
}
