% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDI.r
\name{score_CDI1}
\alias{score_CDI1}
\title{This is a scoring program for the CDI version 1.}
\usage{
score_CDI1(
  meas,
  component = TRUE,
  nacorrect = FALSE,
  navalue = 1,
  nathresh = 3
)
}
\arguments{
\item{navalue}{What should we inpute for missing item scores? 0-2 or NULL. 'fiml' will attempt to use fiml to estimate item scores. NA will retain NA through scoring.}

\item{nathresh}{How many items can be imputed before the subject's score is invalidated?

This is a scoring program for the 26 item Childhood Depression Inventory.
It accepts a data frame with 26 columns corresponding to responses
on the measure (in order).}

\item{data}{n×28 vector of item responses}
}
\description{
\code{score_CDI2} Returns (raw) total symptom score for the CDI.
}
