% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDI.r
\name{score_CDI2}
\alias{score_CDI2}
\title{This is a scoring program for the CDI version 2.}
\usage{
score_CDI2(
  meas,
  component = TRUE,
  nacorrect = FALSE,
  navalue = 1,
  nathresh = 3
)
}
\arguments{
\item{component}{should the component scores be returned?}

\item{navalue}{What should we inpute for missing item scores? 0-2 or NULL. NA will retain NA through scoring.}

\item{nathresh}{How many items can be imputed before the subject's score is invalidated?

This is a scoring program for the 28 item Childhood Depression Inventory.
It accepts a data frame with 28 columns corresponding to responses
on the measure (in order). It returns raw scores only.}

\item{data}{n×28 vector of item responses}
}
\description{
\code{score_CDI2} Returns (raw) total symptom score for the CDI.
}
