% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTQ.R
\name{score_CTQ}
\alias{score_CTQ}
\title{Score the Childhood Trauma Questionnaire}
\usage{
score_CTQ(data, thresh.scores = TRUE)
}
\arguments{
\item{data}{n×28 vector of reported responses.}

\item{thresh.scores}{boolean - program returns threshold scores if TRUE.}
}
\description{
This is a scoring program for the CTQ.
}
\details{
\code{score_CTQ} Returns cluster scores for CTQ subscales, along with dichotomous threshold variables.
}
