% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAST.R
\name{score_DAST}
\alias{score_DAST}
\title{This is a scoring program for the 28 item Drug Abuse Screening Test.
It accepts a data frame with 28 columns corresponding to responses
on the measure (in order).}
\usage{
score_DAST(meas)
}
\description{
\code{score_DAST} Returns a total symptom score.
}
