% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCARED.r
\name{score_SCARED}
\alias{score_SCARED}
\title{This is a scoring program for the SCARED}
\usage{
score_SCARED(meas, nathresh = 3)
}
\arguments{
\item{meas}{data frame or matrix with 41 columns representing SCARED reesponses.}

\item{nathresh}{How many skipped answers are allowed before the subject is marked NA? 0 for no NA handling.}
}
\description{
\code{score_SCARED} Returns total symptom score for the SCARED.
}
