% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdcorr.R
\name{sdcorr}
\alias{sdcorr}
\title{Print a pretty correlation table.}
\usage{
sdcorr(data, digits = 2)
}
\arguments{
\item{data}{a data frame with just numeric data.}

\item{digits}{How many digits to print?}
}
\description{
This function prints a pretty correlation table.
}
