% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starPs.R
\name{starPs}
\alias{starPs}
\title{This function formats '***' style significance strings for a column of p values.}
\usage{
starPs(p, s3 = 0.001, s2 = 0.01, s1 = 0.05)
}
\arguments{
\item{p}{vector of p values.}

\item{s3}{Smallest, ***, significance threshold.}

\item{s2}{Middle, **, significance threshold.}

\item{s1}{Largest, *, significance threshold.}
}
\description{
This function formats '***' style significance strings for a column of p values.
}
