% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttesttable.R
\name{ttesttable}
\alias{ttesttable}
\title{Use t-tests to compare two groups across variables.}
\usage{
ttesttable(varlist, groupvar)
}
\arguments{
\item{varlist}{A list or dataframe of values to compare}

\item{groupvar}{A factor containing the group

Outputs a matrix}
}
\description{
Use t-tests to compare two groups across variables.
}
