% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewSVG.R
\name{ViewSVG}
\alias{ViewSVG}
\title{Opens an SVG or HTML file within an rstudio environment.}
\usage{
ViewSVG(img, use_viewer = TRUE)
}
\arguments{
\item{img}{path/filename of the file to be opened.}

\item{use_viewer}{Use the rstudio viewer (set to FALSE to open a new tab)}
}
\description{
Opens an SVG or HTML file within an rstudio environment.
}
